requirements.txt
playwright
pynacl
requests
